Move component CardRow from 'components/cards' to 'components/index'
