+++ components/index/CardsMonitoring.vue
@@ -28,7 +28,7 @@
 <script lang="ts">
 /* eslint-disable simple-import-sort/sort -- ブラウザでの表示順に合わせて各 card の component を import する */
 import Vue from 'vue'
-import CardRow from '@/components/cards/CardRow.vue'
+import CardRow from '@/components/index/CardRow.vue'
 
 // 検査陽性者の状況
 import ConfirmedCasesDetailsCard from '@/components/cards/ConfirmedCasesDetailsCard.vue'
