+++ components/index/CardsReference.vue
@@ -30,7 +30,7 @@
 <script lang="ts">
 /* eslint-disable simple-import-sort/sort -- ブラウザでの表示順に合わせて各 card の component を import する */
 import Vue from 'vue'
-import CardRow from '@/components/cards/CardRow.vue'
+import CardRow from '@/components/index/CardRow.vue'
 
 // 陽性者の属性
 import ConfirmedCasesAttributesCard from '@/components/cards/ConfirmedCasesAttributesCard.vue'
